Deprecate static methods
